Revert Double

+++ src/wallet/wallet2.cpp
@@ -8652,7 +8652,7 @@ FeePriority wallet2::adjust_priority(FeePriority priority)
       // check if there's a backlog in the tx pool
       const bool use_per_byte_fee = use_fork_rules(HF_VERSION_PER_BYTE_FEE, 0);
       const uint64_t base_fee = get_base_fee(FeePriority::Unimportant);
-      const double fee_level = base_fee * (use_per_byte_fee ? 1 : (12. / 13. / 1024.));
+      const double fee_level = base_fee * (use_per_byte_fee ? 1 : (12/(double)13 / (double)1024));
       const std::vector<std::pair<uint64_t, uint64_t>> blocks = estimate_backlog({std::make_pair(fee_level, fee_level)});
       if (blocks.size() != 1)
       {
